flask/app/requirements.txt
redis
Flask
flask-cors
Authlib
python-dotenv
PyJWT
werkzeug
blinker
jinja2
itsdangerous
cryptography
requests